+++ lib/leptris/xml/sax.rb
@@ -7,6 +7,7 @@ module Leptris
     module SAX
       autoload :Document, "leptris/xml/sax/document"
       autoload :Parser, "leptris/xml/sax/parser"
+      autoload :Recorder, "leptris/xml/sax/recorder"
     end
   end
 end